+++ database/migrations/2021_05_16_120420_create_ps_intakes_table.php
@@ -15,7 +15,7 @@ class CreatePsIntakesTable extends Migration
     {
         Schema::create('ps_intakes', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('referral_source_id')->nullable();
+            $table->unsignedBigInteger('referral_source_id')->nullable();
             $table->date('referral_date');
             $table->date('close_date')->nullable();
             // $table->unsignedBigInteger('elapsed_months_since_intake');
