+++ database/migrations/2021_09_19_062805_create_activities_emergency_types_pivot_table.php
@@ -15,8 +15,8 @@ class CreateActivitiesEmergencyTypesPivotTable extends Migration
     {
         Schema::create('activities_emergency_types', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('activity_id');
-            $table->unsignedInteger('emergency_type_id');
+            $table->unsignedBigInteger('activity_id');
+            $table->unsignedBigInteger('emergency_type_id');
             $table->timestamps();
 
             $table->foreign('activity_id')->references('id')->on('activities')->onDelete('cascade');
